// src/generated/v2/types.gen.ts
export type GetFilteredTeamTasksData = {
    body?: never;
    path: {
        /**
         * Workspace ID
         */
        team_Id: number;
    };
    query?: {
        /**
         * Page to fetch (starts at 0).
         */
        page?: number;
        /**
         * Order by a particular field. By default, tasks are ordered by `created`.\
         * \
         * Options include: `id`, `created`, `updated`, and `due_date`.
         */
        order_by?: string;
        /**
         * Tasks are displayed in reverse order.
         */
        reverse?: boolean;
        /**
         * Include or exclude subtasks. By default, subtasks are excluded.
         */
        subtasks?: boolean;
        /**
         * Filter by Spaces. For example: \
         * \
         * `?space_ids[]=1234&space_ids[]=6789`
         */
        'space_ids[]'?: Array<string>;
        /**
         * Filter by Folders. For example: \
         * \
         * `?project_ids[]=1234&project_ids[]=6789`
         */
        'project_ids[]'?: Array<string>;
        /**
         * Filter by Lists. For example: \
         * \
         * `?list_ids[]=1234&list_ids[]=6789`
         */
        'list_ids[]'?: Array<string>;
        /**
         * Filter by statuses. Use `%20` to represent a space character. To include closed tasks, use the `include_closed` parameter. \
         * \
         * For example: \
         * \
         * `?statuses[]=to%20do&statuses[]=in%20progress`
         */
        'statuses[]'?: Array<string>;
        /**
         * Include or excluse closed tasks. By default, they are excluded.\
         * \
         * To include closed tasks, use `include_closed: true`.
         */
        include_closed?: boolean;
        /**
         * Filter by Assignees using people's ClickUp user id. For example: \
         * \
         * `?assignees[]=1234&assignees[]=5678`
         */
        'assignees[]'?: Array<string>;
        /**
         * Filter by tags. User `%20` to represent a space character. For example: \
         * \
         * `?tags[]=tag1&tags[]=this%20tag`
         */
        'tags[]'?: Array<string>;
        /**
         * Filter by due date greater than Unix time in milliseconds.
         */
        due_date_gt?: number;
        /**
         * Filter by due date less than Unix time in milliseconds.
         */
        due_date_lt?: number;
        /**
         * Filter by date created greater than Unix time in milliseconds.
         */
        date_created_gt?: number;
        /**
         * Filter by date created less than Unix time in milliseconds.
         */
        date_created_lt?: number;
        /**
         * Filter by date updated greater than Unix time in milliseconds.
         */
        date_updated_gt?: number;
        /**
         * Filter by date updated less than Unix time in milliseconds.
         */
        date_updated_lt?: number;
        /**
         * Filter by date done greater than Unix time in milliseconds.
         */
        date_done_gt?: number;
        /**
         * Filter by date done less than Unix time in milliseconds.
         */
        date_done_lt?: number;
        /**
         * Include tasks with specific values in one or more Custom Fields. Custom Relationships are included.\
         * \
         * For example: `?custom_fields=[{"field_id":"abcdefghi12345678","operator":"=","value":"1234"}{"field_id":"jklmnop123456","operator":"<","value":"5"}]`\
         * \
         * Only set Custom Field values display in the `value` property of the `custom_fields` parameter. The `=` operator isn't supported with Label Custom Fields.\
         * \
         * Learn more about [filtering using Custom Fields.](doc:taskfilters)
         */
        custom_fields?: Array<string>;
        /**
         * Include the parent task ID to return subtasks.
         */
        parent?: string;
        /**
         * To return task descriptions in Markdown format, use `?include_markdown_description=true`.
         */
        include_markdown_description?: boolean;
        /**
         * Filter by custom task types. For example: \
         * \
         * `?custom_items[]=0&custom_items[]=1300` \
         * \
         * Including `0` returns tasks. Including `1` returns Milestones. Including any other number returns the custom task type as defined in your Workspace.
         */
        'custom_items[]'?: Array<number>;
    };
    url: '/v2/team/{team_Id}/task';
};
// ...
export type GetTimeEntriesWithinDateRangeData = {
    body?: never;
    headers: {
        'Content-Type': 'application/json';
    };
    path: {
        /**
         * Workspace ID
         */
        team_Id: number;
    };
    query?: {
        /**
         * Unix time in milliseconds
         */
        start_date?: number;
        /**
         * Unix time in milliseconds
         */
        end_date?: number;
        /**
         * Filter by `user_id`. For multiple assignees, separate `user_id` using commas.\
         * \
         * **Example:** `assignee=1234,9876`\
         * \
         * ***Note:** Only Workspace Owners/Admins have access to do this.*
         */
        assignee?: number;
        /**
         * Include task tags in the response for time entries associated with tasks.
         */
        include_task_tags?: boolean;
        /**
         * Include the names of the List, Folder, and Space along with the `list_id`,`folder_id`, and `space_id`.
         */
        include_location_names?: boolean;
        /**
         * Include the history of the approval for each time entry. Adds status changes, notes, and approvers.
         */
        include_approval_history?: boolean;
        /**
         * Include the details of the approval for each time entry. Adds Approver ID, Approved Time, List of Approvers, and Approval Status.
         */
        include_approval_details?: boolean;
        /**
         * Only include time entries associated with tasks in a specific Space.
         */
        space_id?: number;
        /**
         * Only include time entries associated with tasks in a specific Folder.
         */
        folder_id?: number;
        /**
         * Only include time entries associated with tasks in a specific List.
         */
        list_id?: number;
        /**
         * Only include time entries associated with a specific task.
         */
        task_id?: string;
        /**
         * If you want to reference a task by it's custom task id, this value must be `true`.
         */
        custom_task_ids?: boolean;
        /**
         * When the `custom_task_ids` parameter is set to `true`, the Workspace ID must be provided using the `team_id` parameter.
         * \
         * For example: `custom_task_ids=true&team_id=123`.
         */
        team_id?: number;
        /**
         * Include only billable time entries by using a value of `true` or only non-billable time entries by using a value of `false`.\
         * \
         * For example: `?is_billable=true`.
         */
        is_billable?: boolean;
    };
    url: '/v2/team/{team_Id}/time_entries';
};